requirements.txt
Flask==2.1.0
gunicorn==20.1.0
line-bot-sdk==2.4.2
google-cloud-secret-manager==2.16.1